+++ database/migrations/2023_05_26_114537_create_user_categories_table.php
@@ -13,16 +13,21 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('user_categories', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('category_id');
-            $table->timestamps();
+        Schema::create(
+            'user_categories', function (Blueprint $table) {
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('category_id');
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+                if(env('APP_ENV')!='testing') {
+                    $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                    $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+                }
 
-            $table->unique(['user_id', 'category_id']);
-        });
+
+                $table->unique(['user_id', 'category_id']);
+            }
+        );
     }
 
     /**
