+++ database/migrations/2023_05_26_114615_create_user_authors_table.php
@@ -13,17 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('user_authors', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('author_id');
-            $table->timestamps();
-
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('author_id')->references('id')->on('authors')->onDelete('cascade');
-
-            $table->unique(['user_id', 'author_id']);
-        });
-
+        Schema::create(
+            'user_authors', function (Blueprint $table) {
+                $table->unsignedBigInteger('user_id');
+                $table->unsignedBigInteger('author_id');
+                $table->timestamps();
+                if (env('APP_ENV') != 'testing') {
+                    $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                    $table->foreign('author_id')->references('id')->on('authors')->onDelete('cascade');
+                }
+                $table->unique(['user_id', 'author_id']);
+            }
+        );
     }
 
     /**
